Converted numba function signatures for helper and add_noise functions to double-precision

+++ Production/add_noise.py
@@ -1,9 +1,9 @@
 # A nice little function to add random noise to the sensor data.
 
 from random import gauss
-from numba import jit, float32
+from numba import jit, float64
 
 
-@jit(float32)
+@jit(float64)
 def fuzz(value, factor=0.01):
     return gauss(value, factor * value)
